+++ birdwatcher/tests/test_frameprocessing.py
@@ -1,5 +1,7 @@
 import unittest
-from birdwatcher.frames import Frames, FramesColor, framecolor
+import numpy as np
+from numpy.testing import assert_array_equal
+from birdwatcher.frames import Frames, FramesColor, framecolor, framegray
 
 
 class TestFrameColor(unittest.TestCase):
@@ -32,4 +34,21 @@ class TestPeekFrame(unittest.TestCase):
         self.assertEqual(len(framelist), 2) # is the first frame still available?
         self.assertEqual(framelist[1].sum(), self.frame1.sum())
 
+
+class TestFindNonZero(unittest.TestCase):
+
+    def test_grayframe(self):
+        f = framegray(width=640, height=480)
+        f[1, 1] = 1
+        f[1, 2] = 2
+        assert_array_equal(next(Frames([f]).find_nonzero()),
+                           np.array([[1, 1], [2, 1]], dtype='int32'))
+
+    def test_colorframe(self):
+        f = framecolor(width=640, height=480, color=(0, 0, 0))
+        f[1, 1] = (1, 1, 1)
+        f[1, 2] = (1, 0, 0)
+        f[1, 3] = (254, 1, 1) # sums to 0 if uint8 sum
+        assert_array_equal(next(Frames([f]).find_nonzero()),
+                           np.array([[1, 1], [2, 1], [3, 1]], dtype='int32'))