+++ docs/releasenotes.rst
@@ -6,7 +6,10 @@ Master
 
 Commits by Carien Mol and Gabriel Beckers.
 
+- `find_nonzero` methods now can work with color frames
 - `apply_backgroundsegmenter` method on Frames for background segmentation
 - improved logging of processing steps
 - `peek_frames` method for Frames
 - `show` method for Frames, VideoFileStream and CoordinateArrays
+- many improvements docstrings and code consistency
+- added more tests